// client/src/app/core/modules/openapi/model/advert-dto.ts
/**
 * AdBoard API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShortUserDto } from './short-user-dto';
import { ShortCategoryDto } from './short-category-dto';


/**
 * Объявление (подробное описание)
 */
export interface AdvertDto { 
    /**
     * Id.
     */
    id?: string;
    user?: ShortUserDto;
    /**
     * Название.
     */
    name?: string | null;
    /**
     * Описание.
     */
    description?: string | null;
    /**
     * Признак актуальности объявления.
     */
    isActive?: boolean;
    /**
     * ID изображений привязанных к объявлению
     */
    imagesIds?: Array<string> | null;
    /**
     * Цена.
     */
    cost?: number;
    /**
     * Email.
     */
    email?: string | null;
    /**
     * Телефон.
     */
    phone?: string | null;
    /**
     * Геолокация объявления
     */
    location?: string | null;
    /**
     * Дата создания.
     */
    created?: string;
    category?: ShortCategoryDto;
}