// client/src/app/core/modules/openapi/model/advert-search-request-dto.ts
/**
 * AdBoard API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Модель запроса на поиск объявлений.
 */
export interface AdvertSearchRequestDto { 
    /**
     * Строка поиска.
     */
    search?: string | null;
    /**
     * Показывать неактивные объявления
     */
    showNonActive?: boolean | null;
    /**
     * Категория
     */
    category?: string | null;
}